requirements.txt
absl-py==1.0.0
anyio==3.5.0
apturl==0.5.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
dictdiffer==0.8.1
distro==1.4.0
distro-info===0.23ubuntu1
dominate==2.5.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
folder-color-common==0.0.88
future==0.18.2
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
idna==2.8
kazam==1.4.5
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
matplotlib==3.1.2
mediapipe==0.8.9.1
mistletoe==0.7.2
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.4
notion-client==0.9.0
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.20.0
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
pyenchant==2.0.0
PyGetWindow==0.0.9
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
python-apt===2.0.0-ubuntu0.20.04.7-zorin1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-slugify==4.0.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.8
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.1
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
tzlocal==2.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
wadllib==1.3.3
xkit==0.0.0
yapf==0.32.0
zorin-appearance==4.1
zorin-connect==1.2
zorin-exec-guard==1.4
